% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_objects.R
\name{AchievementConfiguration}
\alias{AchievementConfiguration}
\title{Google Play Game Services Publishing API Objects 
The Publishing API for Google Play Game Services.}
\usage{
AchievementConfiguration(achievementType = NULL, draft = NULL, id = NULL,
  initialState = NULL, published = NULL, stepsToUnlock = NULL,
  token = NULL)
}
\arguments{
\item{achievementType}{The type of the achievement}

\item{draft}{The draft data of the achievement}

\item{id}{The ID of the achievement}

\item{initialState}{The initial state of the achievement}

\item{published}{The read-only published data of the achievement}

\item{stepsToUnlock}{Steps to unlock}

\item{token}{The token for this resource}
}
\value{
AchievementConfiguration object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:52:32
filename: /Users/mark/dev/R/autoGoogleAPI/googlegamesConfigurationv1configuration.auto/R/gamesConfiguration_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AchievementConfiguration Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an achievement configuration resource.
}
\seealso{
Other AchievementConfiguration functions: \code{\link{achievementConfigurations.insert}},
  \code{\link{achievementConfigurations.patch}},
  \code{\link{achievementConfigurations.update}}
}
