% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_objects.R
\name{AchievementConfigurationDetail}
\alias{AchievementConfigurationDetail}
\title{AchievementConfigurationDetail Object}
\usage{
AchievementConfigurationDetail(description = NULL, iconUrl = NULL,
  name = NULL, pointValue = NULL, sortRank = NULL)
}
\arguments{
\item{description}{Localized strings for the achievement description}

\item{iconUrl}{The icon url of this achievement}

\item{name}{Localized strings for the achievement name}

\item{pointValue}{Point value for the achievement}

\item{sortRank}{The sort rank of this achievement}
}
\value{
AchievementConfigurationDetail object
}
\description{
AchievementConfigurationDetail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an achievement configuration detail.
}
