% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_objects.R
\name{AchievementConfigurationListResponse}
\alias{AchievementConfigurationListResponse}
\title{AchievementConfigurationListResponse Object}
\usage{
AchievementConfigurationListResponse(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{The achievement configurations}

\item{nextPageToken}{The pagination token for the next page of results}
}
\value{
AchievementConfigurationListResponse object
}
\description{
AchievementConfigurationListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a ListConfigurations response.
}

