% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_functions.R
\name{achievementConfigurations.insert}
\alias{achievementConfigurations.insert}
\title{Insert a new achievement configuration in this application.}
\usage{
achievementConfigurations.insert(AchievementConfiguration, applicationId)
}
\arguments{
\item{AchievementConfiguration}{The \link{AchievementConfiguration} object to pass to this method}

\item{applicationId}{The application ID from the Google Play developer console}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidpublisher
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidpublisher)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services}{Google Documentation}

Other AchievementConfiguration functions: \code{\link{AchievementConfiguration}},
  \code{\link{achievementConfigurations.patch}},
  \code{\link{achievementConfigurations.update}}
}

