% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_functions.R
\name{achievementConfigurations.patch}
\alias{achievementConfigurations.patch}
\title{Update the metadata of the achievement configuration with the given ID. This method supports patch semantics.}
\usage{
achievementConfigurations.patch(AchievementConfiguration, achievementId)
}
\arguments{
\item{AchievementConfiguration}{The \link{AchievementConfiguration} object to pass to this method}

\item{achievementId}{The ID of the achievement used by this method}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidpublisher
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidpublisher)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services}{Google Documentation}

Other AchievementConfiguration functions: \code{\link{AchievementConfiguration}},
  \code{\link{achievementConfigurations.insert}},
  \code{\link{achievementConfigurations.update}}
}
