% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_functions.R
\docType{package}
\name{gamesConfiguration_googleAuthR}
\alias{gamesConfiguration_googleAuthR}
\alias{gamesConfiguration_googleAuthR-package}
\title{Google Play Game Services Publishing API
The Publishing API for Google Play Game Services.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:52:32
filename: /Users/mark/dev/R/autoGoogleAPI/googlegamesConfigurationv1configuration.auto/R/gamesConfiguration_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/androidpublisher
}
}
