% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_objects.R
\name{GamesNumberFormatConfiguration}
\alias{GamesNumberFormatConfiguration}
\title{GamesNumberFormatConfiguration Object}
\usage{
GamesNumberFormatConfiguration(currencyCode = NULL, numDecimalPlaces = NULL,
  numberFormatType = NULL, suffix = NULL)
}
\arguments{
\item{currencyCode}{The curreny code string}

\item{numDecimalPlaces}{The number of decimal places for number}

\item{numberFormatType}{The formatting for the number}

\item{suffix}{An optional suffix for the NUMERIC format type}
}
\value{
GamesNumberFormatConfiguration object
}
\description{
GamesNumberFormatConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a number format resource.
}
