% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_objects.R
\name{ImageConfiguration}
\alias{ImageConfiguration}
\title{ImageConfiguration Object}
\usage{
ImageConfiguration(imageType = NULL, resourceId = NULL, url = NULL)
}
\arguments{
\item{imageType}{The image type for the image}

\item{resourceId}{The resource ID of resource which the image belongs to}

\item{url}{The url for this image}
}
\value{
ImageConfiguration object
}
\description{
ImageConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an image configuration resource.
}
