% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_functions.R
\name{imageConfigurations.upload}
\alias{imageConfigurations.upload}
\title{Uploads an image for a resource with the given ID and image type.}
\usage{
imageConfigurations.upload(resourceId, imageType)
}
\arguments{
\item{resourceId}{The ID of the resource used by this method}

\item{imageType}{Selects which image in a resource for this method}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidpublisher
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidpublisher)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services}{Google Documentation}
}

