% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_objects.R
\name{LeaderboardConfiguration}
\alias{LeaderboardConfiguration}
\title{LeaderboardConfiguration Object}
\usage{
LeaderboardConfiguration(draft = NULL, id = NULL, published = NULL,
  scoreMax = NULL, scoreMin = NULL, scoreOrder = NULL, token = NULL)
}
\arguments{
\item{draft}{The draft data of the leaderboard}

\item{id}{The ID of the leaderboard}

\item{published}{The read-only published data of the leaderboard}

\item{scoreMax}{Maximum score that can be posted to this leaderboard}

\item{scoreMin}{Minimum score that can be posted to this leaderboard}

\item{scoreOrder}{The type of the leaderboard}

\item{token}{The token for this resource}
}
\value{
LeaderboardConfiguration object
}
\description{
LeaderboardConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an leaderboard configuration resource.
}
\seealso{
Other LeaderboardConfiguration functions: \code{\link{leaderboardConfigurations.insert}},
  \code{\link{leaderboardConfigurations.patch}},
  \code{\link{leaderboardConfigurations.update}}
}

