% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_objects.R
\name{LeaderboardConfigurationDetail}
\alias{LeaderboardConfigurationDetail}
\title{LeaderboardConfigurationDetail Object}
\usage{
LeaderboardConfigurationDetail(iconUrl = NULL, name = NULL,
  scoreFormat = NULL, sortRank = NULL)
}
\arguments{
\item{iconUrl}{The icon url of this leaderboard}

\item{name}{Localized strings for the leaderboard name}

\item{scoreFormat}{The score formatting for the leaderboard}

\item{sortRank}{The sort rank of this leaderboard}
}
\value{
LeaderboardConfigurationDetail object
}
\description{
LeaderboardConfigurationDetail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a leaderboard configuration detail.
}

