% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_objects.R
\name{LeaderboardConfigurationListResponse}
\alias{LeaderboardConfigurationListResponse}
\title{LeaderboardConfigurationListResponse Object}
\usage{
LeaderboardConfigurationListResponse(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{The leaderboard configurations}

\item{nextPageToken}{The pagination token for the next page of results}
}
\value{
LeaderboardConfigurationListResponse object
}
\description{
LeaderboardConfigurationListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a ListConfigurations response.
}
