% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_functions.R
\name{leaderboardConfigurations.delete}
\alias{leaderboardConfigurations.delete}
\title{Delete the leaderboard configuration with the given ID.}
\usage{
leaderboardConfigurations.delete(leaderboardId)
}
\arguments{
\item{leaderboardId}{The ID of the leaderboard}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidpublisher
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidpublisher)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services}{Google Documentation}
}
