% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_functions.R
\name{leaderboardConfigurations.insert}
\alias{leaderboardConfigurations.insert}
\title{Insert a new leaderboard configuration in this application.}
\usage{
leaderboardConfigurations.insert(LeaderboardConfiguration, applicationId)
}
\arguments{
\item{LeaderboardConfiguration}{The \link{LeaderboardConfiguration} object to pass to this method}

\item{applicationId}{The application ID from the Google Play developer console}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidpublisher
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidpublisher)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services}{Google Documentation}

Other LeaderboardConfiguration functions: \code{\link{LeaderboardConfiguration}},
  \code{\link{leaderboardConfigurations.patch}},
  \code{\link{leaderboardConfigurations.update}}
}

