% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_functions.R
\name{leaderboardConfigurations.list}
\alias{leaderboardConfigurations.list}
\title{Returns a list of the leaderboard configurations in this application.}
\usage{
leaderboardConfigurations.list(applicationId, maxResults = NULL,
  pageToken = NULL)
}
\arguments{
\item{applicationId}{The application ID from the Google Play developer console}

\item{maxResults}{The maximum number of resource configurations to return in the response, used for paging}

\item{pageToken}{The token returned by the previous request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidpublisher
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidpublisher)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services}{Google Documentation}
}
