% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_functions.R
\name{leaderboardConfigurations.update}
\alias{leaderboardConfigurations.update}
\title{Update the metadata of the leaderboard configuration with the given ID.}
\usage{
leaderboardConfigurations.update(LeaderboardConfiguration, leaderboardId)
}
\arguments{
\item{LeaderboardConfiguration}{The \link{LeaderboardConfiguration} object to pass to this method}

\item{leaderboardId}{The ID of the leaderboard}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/androidpublisher
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/androidpublisher)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services}{Google Documentation}

Other LeaderboardConfiguration functions: \code{\link{LeaderboardConfiguration}},
  \code{\link{leaderboardConfigurations.insert}},
  \code{\link{leaderboardConfigurations.patch}}
}

