% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_objects.R
\name{LocalizedString}
\alias{LocalizedString}
\title{LocalizedString Object}
\usage{
LocalizedString(locale = NULL, value = NULL)
}
\arguments{
\item{locale}{The locale string}

\item{value}{The string value}
}
\value{
LocalizedString object
}
\description{
LocalizedString Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a localized string resource.
}

