% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesConfiguration_objects.R
\name{LocalizedStringBundle}
\alias{LocalizedStringBundle}
\title{LocalizedStringBundle Object}
\usage{
LocalizedStringBundle(translations = NULL)
}
\arguments{
\item{translations}{The locale strings}
}
\value{
LocalizedStringBundle object
}
\description{
LocalizedStringBundle Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a localized string bundle resource.
}
